Rename config.py to database.py
