+++ src/lifedrain.py
@@ -3,7 +3,7 @@ Copyright (c) Yutsuten <https://github.com/Yutsuten>. Licensed under AGPL-3.0.
 See the LICENCE file in the repository root for full licence text.
 """
 
-from .config import GlobalConf, DeckConf
+from .database import GlobalConf, DeckConf
 from .deck_manager import DeckManager
 from .decorators import must_be_enabled
 from . import settings
